# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LexAI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      parameters:
        - in: query
          name: matterId
          required: false
          schema:
            type: integer
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [draft, review, signed]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a contract manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/draft:
    post:
      operationId: draftContract
      tags: [contracts]
      summary: AI-generate a contract from a prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftContractInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    patch:
      operationId: updateContract
      tags: [contracts]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    delete:
      operationId: deleteContract
      tags: [contracts]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
  /contracts/{id}/analyze:
    post:
      operationId: analyzeContract
      tags: [contracts]
      summary: AI risk-score and flag clauses
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
...
  /deadlines:
    get:
      operationId: listDeadlines
      tags: [deadlines]
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [upcoming, completed, missed]
        - in: query
          name: matterId
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deadline"
    post:
      operationId: createDeadline
      tags: [deadlines]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeadlineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deadline"
...
  /deadlines/{id}:
    patch:
      operationId: updateDeadline
      tags: [deadlines]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeadlineInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deadline"
    delete:
      operationId: deleteDeadline
      tags: [deadlines]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
...
  /briefs:
    get:
      operationId: listBriefs
      tags: [briefs]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brief"
    post:
      operationId: createBrief
      tags: [briefs]
      summary: Generate a plain-English client brief from legalese
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBriefInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brief"
  /briefs/{id}:
    get:
      operationId: getBrief
      tags: [briefs]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brief"
    delete:
      operationId: deleteBrief
      tags: [briefs]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    Jurisdiction:
      type: string
      enum: [india, uk, us]
...
    ContractType:
      type: string
      enum: [nda, msa, employment, lease, spa, services, partnership, licensing, consulting, other]
    RiskFlag:
      type: object
      required: [clause, severity, explanation]
      properties:
        clause: { type: string }
        severity:
          type: string
          enum: [low, medium, high, critical]
        explanation: { type: string }
    Contract:
      type: object
      required:
        - id
        - title
        - type
        - jurisdiction
        - status
        - content
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        matterId:
          type: ["integer", "null"]
        title: { type: string }
        type:
          $ref: "#/components/schemas/ContractType"
        jurisdiction:
          $ref: "#/components/schemas/Jurisdiction"
        status:
          type: string
          enum: [draft, review, signed]
        counterparty:
          type: ["string", "null"]
        content: { type: string }
        riskScore:
          type: ["integer", "null"]
        riskFlags:
          type: array
          items:
            $ref: "#/components/schemas/RiskFlag"
        summary:
          type: ["string", "null"]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateContractInput:
      type: object
      required: [title, type, jurisdiction, content]
      properties:
        matterId:
          type: ["integer", "null"]
        title: { type: string }
        type:
          $ref: "#/components/schemas/ContractType"
        jurisdiction:
          $ref: "#/components/schemas/Jurisdiction"
        status:
          type: string
          enum: [draft, review, signed]
        counterparty:
          type: ["string", "null"]
        content: { type: string }
    UpdateContractInput:
      type: object
      properties:
        title: { type: string }
        type:
          $ref: "#/components/schemas/ContractType"
        jurisdiction:
          $ref: "#/components/schemas/Jurisdiction"
        status:
          type: string
          enum: [draft, review, signed]
        counterparty:
          type: ["string", "null"]
        content: { type: string }
    DraftContractInput:
      type: object
      required: [title, type, jurisdiction, prompt]
      properties:
        matterId:
          type: ["integer", "null"]
        title: { type: string }
        type:
          $ref: "#/components/schemas/ContractType"
        jurisdiction:
          $ref: "#/components/schemas/Jurisdiction"
        counterparty:
          type: ["string", "null"]
        prompt:
          type: string
          description: Plain-English description of the deal
...
    Deadline:
      type: object
      required: [id, title, dueDate, type, status, priority, createdAt]
      properties:
        id: { type: integer }
        matterId:
          type: ["integer", "null"]
        matterName:
          type: ["string", "null"]
        title: { type: string }
        description:
          type: ["string", "null"]
        dueDate: { type: string, format: date-time }
        type:
          type: string
          enum: [filing, court_date, statute, internal, client_meeting]
        status:
          type: string
          enum: [upcoming, completed, missed]
        priority:
          type: string
          enum: [low, medium, high]
        createdAt: { type: string, format: date-time }
    CreateDeadlineInput:
      type: object
      required: [title, dueDate, type, priority]
      properties:
        matterId:
          type: ["integer", "null"]
        title: { type: string }
        description:
          type: ["string", "null"]
        dueDate: { type: string, format: date-time }
        type:
          type: string
          enum: [filing, court_date, statute, internal, client_meeting]
        priority:
          type: string
          enum: [low, medium, high]
    UpdateDeadlineInput:
      type: object
      properties:
        title: { type: string }
        description:
          type: ["string", "null"]
        dueDate: { type: string, format: date-time }
        type:
          type: string
          enum: [filing, court_date, statute, internal, client_meeting]
        status:
          type: string
          enum: [upcoming, completed, missed]
        priority:
          type: string
          enum: [low, medium, high]
...
    Brief:
      type: object
      required: [id, title, originalText, plainEnglish, complexity, createdAt]
      properties:
        id: { type: integer }
        matterId:
          type: ["integer", "null"]
        title: { type: string }
        originalText: { type: string }
        plainEnglish: { type: string }
        keyPoints:
          type: array
          items: { type: string }
        complexity:
          type: string
          enum: [basic, standard, detailed]
        createdAt: { type: string, format: date-time }
    CreateBriefInput:
      type: object
      required: [title, originalText, complexity]
      properties:
        matterId:
          type: ["integer", "null"]
        title: { type: string }
        originalText: { type: string }
        complexity:
          type: string
          enum: [basic, standard, detailed]
...