# lib/api-zod/src/generated/types/brief.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LexAI API
 * OpenAPI spec version: 0.1.0
 */
import type { BriefComplexity } from "./briefComplexity";

export interface Brief {
  id: number;
  /** @nullable */
  matterId?: number | null;
  title: string;
  originalText: string;
  plainEnglish: string;
  keyPoints?: string[];
  complexity: BriefComplexity;
  createdAt: Date;
}